Improve PHP 8.5 compatibility

+++ includes/system/versioned/1.01.00.00/product_searcher.php
@@ -111,18 +111,18 @@ EOSQL
           , (int)$country_id, (int)$zone_id);
 
           if ($GLOBALS['pfrom'] > 0) {
-            $custom['where'] .= " AND (IF(s.status, s.specials_new_products_price, p.products_price) * (1.0 + (tax.tax_rate / 100) ) >= " . (double)$GLOBALS['pfrom'] . ")";
+            $custom['where'] .= " AND (IF(s.status, s.specials_new_products_price, p.products_price) * (1.0 + (tax.tax_rate / 100) ) >= " . (float)$GLOBALS['pfrom'] . ")";
           }
           if ($GLOBALS['pto'] > 0) {
-            $custom['where'] .= " AND (IF(s.status, s.specials_new_products_price, p.products_price) * (1.0 + (tax.tax_rate / 100) ) <= " . (double)$GLOBALS['pto'] . ")";
+            $custom['where'] .= " AND (IF(s.status, s.specials_new_products_price, p.products_price) * (1.0 + (tax.tax_rate / 100) ) <= " . (float)$GLOBALS['pto'] . ")";
           }
         }
       } else {
         if ($GLOBALS['pfrom'] > 0) {
-          $custom['where'] .= " AND IF(s.status, s.specials_new_products_price, p.products_price) >= " . (double)$GLOBALS['pfrom'];
+          $custom['where'] .= " AND IF(s.status, s.specials_new_products_price, p.products_price) >= " . (float)$GLOBALS['pfrom'];
         }
         if ($GLOBALS['pto'] > 0) {
-          $custom['where'] .= " AND IF(s.status, s.specials_new_products_price, p.products_price) <= " . (double)$GLOBALS['pto'];
+          $custom['where'] .= " AND IF(s.status, s.specials_new_products_price, p.products_price) <= " . (float)$GLOBALS['pto'];
         }
       }
